modernize: change C++ standard from C++98 to C++17
This repository was started in January of 2011, a time when it
still was not clear when C++11 would be finalized by the C++
standards committee. Some compilers even referred to it as C++0x,
because it was believed earlier that the new standard would be
finished somewhen in the 200x years. Alas, it was not. So that is
why this repository contains code that is still written in C++98
style. Now, ten years later, that should (or: has to?) change.

We all want to use new, nice features of the newer standards, and
C++17 is now more than three years old, so any reasonably recent
C++ compiler should support that. As far as I know C++17 language
features are fully supported by (among others):

* GCC 7 (or newer)
* Clang 5 (or newer)
* VS 2017 15.7 (or newer)

Some C++ library features may require newer versions, though.

Since even Debian 10 stable (codename "buster") has GCC 8.3.0 and
Clang 7.0 - and Debian is not exactly known for being a bleeding
edge distribution - any reasonably up to date distribution should
have a suitable compiler for C++17. If your distribution has no
such compiler, you should probably consider updating to a more
recent distribution.

The future is now, old compilers! :)