ci: skip separate syntax check on Clang 8, Clang 9, GCC 8, GCC 9
The syntax check usually takes twice as long as the actual
compiling, because not all files are used to produce binaries.
The syntax check is still performed on Clang 10 and GCC 10, so we
should be on the safe side. Clang <= 9 and GCC <= 9 just do the
build and tests from now onwards.