add fix for failed MinGW detection on GitHub Actions
Somehow the CMake detection for the MinGW compiler stopped
working on GitHub on 27th / 28th November 2021. Although it looks
as if the versions of the packages (CMake, compiler, libraries)
used by the GHA build stayed the same, GHA's CMake no longer
properly detects a MinGW compiler when cross-compiling from Linux
to Windows (32 bit).

Reasons are unclear, but it appeared to fail after to the GitHub
outage on that day was resolved:
<https://www.githubstatus.com/incidents/r5qrpp2f5fc0>