bsa-cli: adjust header text of file-metadata command
Instead of "file size" it does now say "block size". This name is
more accurate. The size value stored in a BSA is not necessarily
the size of the file, i. e. when the file data is compressed or
when there are embedded file names in the BSA. However, it is
always the size of the data block that contains the file data.