tests: run test for umlauts in paths of BSA files on Linux systems, too
Initially, the test was only run on Windows, because that is
where umlauts or special characters were problematic. Linux just
uses UTF-8 and everything is fine, so technically it does not
need the test. But to make sure it stays that way, the test is
now run on non-Windows systems, too.