cell_translator: indicate proper encoding in cells.xml
The actual encoding of the file has not changed. The file is and
was encoded as Windows-1252, but the declared encoding in the XML
encoding attribute was UTF-8, and that was simply giving the
wrong idea.

Besides that, the file is not 100 % proper XML, because the
ampersand character is not "escaped" by using the proper XML
entity for it.

[ci skip]