MWTP: adjust RegionRecord to remember size of WEAT sub record
Previously, there were two save methods - the normal one
(saveToStream()) and one with an extra parameter to explicitly
set whether the weather data of the RegionRecord uses the old
Morrowind style (only 8 bytes long) or the newer Bloodmoon style
(10 bytes long). While this may work in some cases, it is
cumbersome and the user has to remember which style to use when
the record is saved. Now the record itself remembers that and any
record loaded from a stream will use the proper style when being
saved to a stream. That's much easier to use.